# docs/openapi.yaml
openapi: 3.1.0
info:
  title: HotelMate API
  version: 1.0.0
  description: Production-ready tenant-aware hotel operations, revenue reporting, approved-knowledge conversations, security audit trails, and live events. Every HTTP response includes an X-Request-ID correlation header.
servers:
  - url: /api/v1
security: []
paths:
...
  /staff/reports/operations:
    get:
      summary: Read tenant-scoped operational and revenue reporting
      description: Available to hotel administrators and operations managers. Dates are interpreted in the configured hotel timezone and the inclusive range may not exceed 31 days.
      security: [{ staffBearer: [] }]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200':
          description: Server-calculated request, revenue, occupancy, handoff, and security metrics
          content:
            application/json:
              schema:
                type: object
                required: [report]
                properties:
                  report: { $ref: '#/components/schemas/OperationalReport' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /staff/audit-logs:
    get:
      summary: Read the tenant security and mutation audit trail
      description: Restricted to primary and secondary hotel administrators. Identity numbers, credentials, and message bodies are never included.
      security: [{ staffBearer: [] }]
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: action
          schema: { type: string, maxLength: 80 }
        - in: query
          name: outcome
          schema: { type: string, enum: [success, failure] }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0, default: 0 }
      responses:
        '200':
          description: Reverse-chronological tenant audit page
          content:
            application/json:
              schema:
                type: object
                required: [audit]
                properties:
                  audit: { $ref: '#/components/schemas/AuditPage' }
        '400': { $ref: '#/components/responses/ValidationFailed' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
...
  securitySchemes:
    staffBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with the hotelmate:staff audience.
    guestBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT with the hotelmate:guest audience and an active stay.
  schemas:
...
    ReportSummary:
      type: object
      required: [requestsCreated, openRequests, completedRequests, averageFulfillmentMinutes, paidOrders, orderedRevenueCents, recognizedRevenueCents, activeRooms, totalRooms, handedOffConversations, pendingKnowledge, failedSecurityEvents]
      properties:
        requestsCreated: { type: integer, format: int64 }
        openRequests: { type: integer, format: int64 }
        completedRequests: { type: integer, format: int64 }
        averageFulfillmentMinutes: { type: integer, format: int64 }
        paidOrders: { type: integer, format: int64 }
        orderedRevenueCents: { type: integer, format: int64, description: Integer IRR amount; compatibility field name retained. }
        recognizedRevenueCents: { type: integer, format: int64, description: Completed-order revenue in integer IRR. }
        activeRooms: { type: integer, format: int64 }
        totalRooms: { type: integer, format: int64 }
        handedOffConversations: { type: integer, format: int64 }
        pendingKnowledge: { type: integer, format: int64 }
        failedSecurityEvents: { type: integer, format: int64 }
    OperationalReport:
      type: object
      required: [from, to, timezone, currency, summary, daily, byService]
      properties:
        from: { type: string, format: date-time }
        to: { type: string, format: date-time }
        timezone: { type: string }
        currency: { const: IRR }
        summary: { $ref: '#/components/schemas/ReportSummary' }
        daily:
          type: array
          items:
            type: object
            required: [date, requests, completed, revenueCents]
            properties:
              date: { type: string, format: date }
              requests: { type: integer, format: int64 }
              completed: { type: integer, format: int64 }
              revenueCents: { type: integer, format: int64 }
        byService:
          type: array
          items:
            type: object
            required: [serviceId, serviceName, orders, quantity, orderedRevenueCents, recognizedRevenueCents]
            properties:
              serviceId: { type: string, format: uuid }
              serviceName: { type: string }
              orders: { type: integer, format: int64 }
              quantity: { type: integer, format: int64 }
              orderedRevenueCents: { type: integer, format: int64 }
              recognizedRevenueCents: { type: integer, format: int64 }
    AuditLog:
      type: object
      required: [id, createdAt, requestId, actorType, action, outcome, ipAddress, metadata]
      properties:
        id: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        requestId: { type: string }
        actorId: { type: string, format: uuid }
        actorType: { type: string }
        action: { type: string }
        outcome: { type: string, enum: [success, failure] }
        ipAddress: { type: string }
        metadata: { type: object, additionalProperties: true }
    AuditPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/AuditLog' } }
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
  responses:
...
    Forbidden:
      description: Role is not permitted
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationFailed:
      description: Input validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many requests for the applicable API, mutation, authentication, or onboarding window
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds before retrying.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }